Refactor screenshot() into util.c
