+++ src/util.c
@@ -3,8 +3,12 @@
 #include <stdlib.h>
 #include <errno.h>
 #include "lodepng.h"
+#include "miniz.h"
 #include "matrix.h"
 #include "util.h"
+#ifdef __EMSCRIPTEN__
+#include <emscripten.h>
+#endif
 
 int rand_int(int n) {
     int result;
@@ -320,4 +324,54 @@ int wrap(const char *input, int max_width, char *output, int max_length) {
     }
     free(text);
     return line_number;
+}
+
+void screenshot(int width, int height) {
+#ifdef __EMSCRIPTEN__
+    EM_ASM(
+        var url = document.getElementsByTagName("canvas")[0].toDataURL();
+        var a = document.createElement("a");
+        var timestamp = new Date().toISOString();
+        a.setAttribute("download", "screenshot-netcraft-" + timestamp + ".png");
+        a.setAttribute("href", url);
+        a.click();
+    );
+#else
+    int channels = 4;
+    int size = channels * width * height;
+    GLubyte *pixels = malloc(size);
+    if (pixels) {
+        glReadPixels(0, 0, width, height, GL_RGBA, GL_UNSIGNED_BYTE, pixels);
+        printf("read pixels to %p\n", pixels);
+
+        size_t png_size = 0;
+        int level = MZ_DEFAULT_LEVEL;
+        void *png = tdefl_write_image_to_png_file_in_memory_ex(pixels,
+                width, height, channels, &png_size, level, MZ_TRUE);
+        if (!png) {
+            printf("failed to write png of screenshot\n");
+        } else {
+            char filename[256];
+            char timestamp[32];
+            time_t now;
+            time(&now);
+            strftime(timestamp, sizeof(timestamp), "%FT%TZ", gmtime(&now));
+            snprintf(filename, sizeof(filename), "screenshot-netcraft-%s.png", timestamp);
+
+            FILE *fp = fopen(filename, "wb");
+            if (fp) {
+                size_t wrote = fwrite(png, 1, png_size, fp);
+                fclose(fp);
+                printf("Saved screenshot to %s\n", filename);
+            } else {
+                printf("failed to open %s for writing\n", filename);
+            }
+            mz_free(png);
+        }
+
+        free(pixels);
+    } else {
+        printf("malloc fail for screenshot (%d bytes)\n", size);
+    }
+#endif
 }
